test(core): keep the determinism test inside its owning tree
The AC19 twice-run test read this repository's roster fixtures from
inside the close-work pack suite, so `lint-pack-test-boundary.py`
correctly rejected it: a pack test may not reach above its owning pack.
Relocated to `tests/roster/test_close_work_extraction_and_immediate_disposition.py`,
which already anchors at the repository root and owns those fixtures, and
reuses that file's `_fixture`/`_inputs` helpers rather than re-deriving
the paths.

Spec: docs/specs/close-work-extraction-and-immediate-disposition/spec.md
